working version

+++ main.cpp
@@ -57,8 +57,7 @@ int main(int argc, char *argv[])
 
     }
 
-    tcpSocketServer *ts = new tcpSocketServer(NULL,IP,PORT);
-
+  tcpSocketServer *ts = new tcpSocketServer(NULL,IP,PORT.toInt());
 
     return a.exec();
 }
